chore: upgrade config delete service

+++ src/services/meal/DeleteMealService.ts
@@ -1,5 +1,6 @@
 import { MealRepository } from '../../repositories/MealRepository';
 import { AppDataSource } from '../../ormconfig';
+import { Meal } from '../../entities/Meal';
 
 export class DeleteMealService {
   private mealRepository: MealRepository;
@@ -8,7 +9,18 @@ export class DeleteMealService {
     this.mealRepository = new MealRepository(AppDataSource);
   }
 
-  async execute(mealId: number, userId: number): Promise<boolean> {
-    return this.mealRepository.deleteMeal(mealId, userId);
+  async execute(mealId: number, userId: number): Promise<Meal | null> {
+    const meal = await this.mealRepository.findOneBy({
+      id: mealId,
+      user: { id: userId },
+    });
+
+    if (!meal) {
+      throw new Error('Meal not found or not authorized to delete this meal.');
+    }
+
+    await this.mealRepository.deleteMeal(mealId, userId);
+
+    return meal;
   }
 }
